added html templating feature. Now can generate html file with embedded graph
