--- pdf_build_resources/rendered_html
@@ -1,31 +1,0 @@
-<!DOCTYPE HTML>
-<!-- Template for use with jinja2 -->
-<html>
-    <p><img src="Test_logo.jpg" alt="logo" /></p>
-    <p>&nbsp;</p>
-    <p>Auto generated report:</p>
-    <p>&nbsp;</p>
-    <p>&nbsp;</p>
-
-
-    <!-- start for loop for figures -->
-    
-        <p><img src="plots/plot1.jpg" alt="plot1.jpg" /></p>
-        <p>&nbsp;</p>
-        <p>&nbsp;</p>
-    
-        <p><img src="plots/plot2.jpg" alt="plot2.jpg" /></p>
-        <p>&nbsp;</p>
-        <p>&nbsp;</p>
-    
-        <p><img src="plots/plot4.jpg" alt="plot4.jpg" /></p>
-        <p>&nbsp;</p>
-        <p>&nbsp;</p>
-    
-        <p><img src="plots/plot3.jpg" alt="plot3.jpg" /></p>
-        <p>&nbsp;</p>
-        <p>&nbsp;</p>
-    
-
-
-</html>
